Add the prefix "Ac" to all ActivityPub classes.
This means that "AcDelete" doesn't create a "delete"
property on related objects which masks the object's delete method.

Also renamed d_k.models.thing to d_k.models.acobject.
(The other modules in models/ weren't renamed because the rename to
"acobject" was only necessary to allow us to import it.)

This has required us to nuke the existing migrations and start again.

Closes #9.