create() fills in as many fields as it can when it creates an AcObject, then sets the others by subscripting.
"As many fields as it can" means
  - the id field
  - any field whose name begins with "f_".

Previously, create() filled in as few fields as it could
(typically only the id, if supplied), then set the others by subscripting.

This change increases efficiency, but also allows the save() method
of AcObject and its subclasses to take account of the intial values of
the f_* fields.