AcActor's f_publicKey renamed to publicKey.
"f_publicKey" would mean it can be set automatically. Because it's
typically a JSON object, it can't be represented with a simple type,
so it can't be set automatically; it must be represented in ThingField.

This didn't make a practical difference before, but since commit 6ce087
create() has honoured the "f_*" convention.