Remote updates are not partial. In other words, fields which are not mentioned in a remote update should get deleted.
This is rarely important in practice, but we should support it.

This commit fixes that problem, but breaks the tests. Thus it
lives in a side branch until someone gets around to implementing
it.

The specific breakage is that AcObject has no __delitem__() method.
It probably should have one for completeness, even apart from this
particular issue.

There should be tests specifically for this problem in test_update.