The "serial" field of bowler_pub's AcItem moves up to AcObject; all objects, remote and local, now have serial numbers.
This is partly to make it easier for trilby_api to provide
decimal "id" numbers, and partly so that status URLs can contain
monotonically increasing decimal numbers. This last part
has not yet been implemented; status URLs still contain the
hex ID of the status.

It's possible that we can simplify this design a bit;
we should think about refactoring.

The regexp for local hex numbers is renamed to LOCAL_NUMBER_REGEXP
because it used to contain the word "SERIAL", which was misleading.

trilby_api now returns the "serial" field as "id"; see above.

Tests updated.