Test utilities involving local users have moved to trilby_api.
This is a bit complicated. We used to have two functions called
create_local_person(): one of them was in bowler_pub and the other
was in trilby_api. The bowler_pub one worked better because it
defaulted to using test keys rather than expecting the system
to make new ones, which is slow and wasteful.

Therefore, we have moved the c_l_p() from bowler into trilby,
and thrown the trilby one out.

While we're at it, we have also moved create_local_status and
create_local_like from bowler into trilby, and deleted trilby's
create_local_note.

This makes everything faster and easier.