Several view fixes in bowler:
CollectionView.get() correctly passes username and listname through to activity_get().
This fixes a regression.

CollectionView.get() dereferences the result of activity_get() if it's an ActivityResponse.
This should have been part of commit d02042f2 but I missed it.
This fixes a regression.

CollectionView.get() uses len() to get the length of the result of activity_get(), and
not the result's count() method. This allows the result to be a list as well as a queryset.

CollectionView.activity_get()'s logging is slightly tidier.